Working bf16

+++ src/operators/fully-connected-nc.c
@@ -151,7 +151,7 @@ static enum xnn_status create_fully_connected_nc(
   const bool block_wise = (block_size != 0);
   if (block_wise) {
     num_blocks = input_channels / block_size;
-    block_scale_bytes += num_blocks * sizeof(float);
+    block_scale_bytes += num_blocks * sizeof(uint16_t);
   }
 
   const size_t weights_stride =
@@ -264,9 +264,9 @@ static enum xnn_status create_fully_connected_nc(
         void* weights_start = (void*) ((uintptr_t) weights_ptr +
           gemm_config->nr * (sizeof(float) + (block_size * sizeof(int8_t) / 2)));
 
-        const size_t block_stride = /*weights*/block_size / 2 + sizeof(float);
+        const size_t block_stride = /*weights*/block_size / 2 + sizeof(uint16_t);
         const size_t weights_stride = /*weights*/k_stride * sizeof(int8_t) +
-            /*scales*/num_blocks * sizeof(float) +
+            /*scales*/num_blocks * sizeof(uint16_t) +
             /*ksum*/sizeof(float) +
             /*bias*/sizeof(float);
 
@@ -628,7 +628,7 @@ enum xnn_status xnn_create_fully_connected_nc_qd8_f16_qb4w(
     input_stride, output_stride,
     kernel, bias, flags,
     /*block_size=*/block_size,
-    /*extra_bl_bytes=*/sizeof(float),
+    /*extra_bl_bytes=*/sizeof(uint16_t),
     /*blockwise_kernel_scale_params=*/kernel_scale,
     /*log2_input_element_size=*/XNN_LOG2_SIZEOF_HALF,
     /*log2_filter_element_size=*/XNN_LOG2_SIZEOF_HALF,
@@ -947,7 +947,7 @@ enum xnn_status xnn_create_fully_connected_nc_qd8_f32_qb4w(
     input_stride, output_stride,
     kernel, bias, flags,
     /*block_size=*/block_size,
-    /*extra_bl_bytes=*/sizeof(float),
+    /*extra_bl_bytes=*/sizeof(uint16_t),
     /*blockwise_kernel_scale_params=*/kernel_scale,
     /*log2_input_element_size=*/XNN_LOG2_SIZEOF_INT8_T,
     /*log2_filter_element_size=*/XNN_LOG2_SIZEOF_UINT8_T,
